yjhv-guys-Avito/requirements.txt
flask
flask_cors
torch
torchvision
pillow
transformers